@@ -22,7 +22,7 @@ import 'utils/location_utils.dart';
 class LocationPicker extends StatefulWidget {
   LocationPicker(
     this.apiKey, {
-    Key key,
+    Key? key,
     this.initialCenter,
     this.initialZoom,
     this.requiredGPS,
@@ -51,35 +51,35 @@ class LocationPicker extends StatefulWidget {
 
   final String apiKey;
 
-  final LatLng initialCenter;
-  final double initialZoom;
-  final List<String> countries;
+  final LatLng? initialCenter;
+  final double? initialZoom;
+  final List<String>? countries;
 
-  final bool requiredGPS;
-  final bool myLocationButtonEnabled;
-  final bool layersButtonEnabled;
-  final bool automaticallyAnimateToCurrentLocation;
+  final bool? requiredGPS;
+  final bool? myLocationButtonEnabled;
+  final bool? layersButtonEnabled;
+  final bool? automaticallyAnimateToCurrentLocation;
 
-  final String mapStylePath;
+  final String? mapStylePath;
 
-  final Color appBarColor;
-  final Color layersIconColor;
-  final Color layersButtonColor;
-  final Color myLocationIconColor;
-  final Color myLocationButtonColor;
-  final Color selectButtonColor;
-  final Color selectButtonFontColor;
-  final BoxDecoration searchBarBoxDecoration;
-  final String hintText;
-  final Widget resultCardConfirmIcon;
-  final Alignment resultCardAlignment;
-  final Decoration resultCardDecoration;
-  final EdgeInsets resultCardPadding;
+  final Color? appBarColor;
+  final Color? layersIconColor;
+  final Color? layersButtonColor;
+  final Color? myLocationIconColor;
+  final Color? myLocationButtonColor;
+  final Color? selectButtonColor;
+  final Color? selectButtonFontColor;
+  final BoxDecoration? searchBarBoxDecoration;
+  final String? hintText;
+  final Widget? resultCardConfirmIcon;
+  final Alignment? resultCardAlignment;
+  final Decoration? resultCardDecoration;
+  final EdgeInsets? resultCardPadding;
 
-  final String language;
-  final String selectButtonText;
+  final String? language;
+  final String? selectButtonText;
 
-  final LocationAccuracy desiredAccuracy;
+  final LocationAccuracy? desiredAccuracy;
 
   @override
   LocationPickerState createState() => LocationPickerState();
@@ -87,12 +87,12 @@ class LocationPicker extends StatefulWidget {
 
 class LocationPickerState extends State<LocationPicker> {
   /// Result returned after user completes selection
-  LocationResult locationResult;
+  late LocationResult locationResult;
 
-  LatLng currentLocation;
+  LatLng? currentLocation;
 
   /// Overlay to display autocomplete suggestions
-  OverlayEntry overlayEntry;
+  OverlayEntry? overlayEntry;
 
   List<NearbyPlace> nearbyPlaces = [];
 
@@ -110,7 +110,7 @@ class LocationPickerState extends State<LocationPicker> {
   /// Hides the autocomplete overlay
   void clearOverlay() {
     if (overlayEntry != null) {
-      overlayEntry.remove();
+      overlayEntry!.remove();
       overlayEntry = null;
     }
   }
@@ -128,14 +128,14 @@ class LocationPickerState extends State<LocationPicker> {
 
     if (place.length < 1) return;
 
-    final RenderBox renderBox = context.findRenderObject();
+    final RenderBox renderBox = context.findRenderObject() as RenderBox;
     Size size = renderBox.size;
 
-    final RenderBox appBarBox = appBarKey.currentContext.findRenderObject();
+    final RenderBox? appBarBox = appBarKey.currentContext!.findRenderObject() as RenderBox?;
 
     overlayEntry = OverlayEntry(
       builder: (context) => Positioned(
-        top: appBarBox.size.height,
+        top: appBarBox!.size.height,
         width: size.width,
         child: Padding(
           padding: const EdgeInsets.only(left: 16.0, right: 16.0),
@@ -165,7 +165,7 @@ class LocationPickerState extends State<LocationPicker> {
       ),
     );
 
-    Overlay.of(context).insert(overlayEntry);
+    Overlay.of(context)!.insert(overlayEntry!);
 
     autoCompleteSearch(place);
   }
@@ -178,7 +178,7 @@ class LocationPickerState extends State<LocationPicker> {
 
     // Currently, you can use components to filter by up to 5 countries. from https://developers.google.com/places/web-service/autocomplete
     String regionParam = countries?.isNotEmpty == true
-        ? "&components=country:${countries.sublist(0, min(countries.length, 5)).join('|country:')}"
+        ? "&components=country:${countries!.sublist(0, min(countries.length, 5)).join('|country:')}"
         : "";
 
     var endpoint =
@@ -188,12 +188,12 @@ class LocationPickerState extends State<LocationPicker> {
             "language=${widget.language}";
 
     if (currentLocation != null) {
-      endpoint += "&location=${currentLocation.latitude}," +
-          "${currentLocation.longitude}&radius=1000";
+      endpoint += "&location=${currentLocation!.latitude}," +
+          "${currentLocation!.longitude}&radius=1000";
     }
 
     LocationUtils.getAppHeaders()
-        .then((headers) => http.get(Uri.parse(endpoint), headers: headers))
+        .then((headers) => http.get(Uri.parse(endpoint), headers: headers as Map<String, String>?))
         .then((response) {
       if (response.statusCode == 200) {
         Map<String, dynamic> data = jsonDecode(response.body);
@@ -233,7 +233,7 @@ class LocationPickerState extends State<LocationPicker> {
   /// To navigate to the selected place from the autocomplete list to the map,
   /// the lat,lng is required. This method fetches the lat,lng of the place and
   /// proceeds to moving the map to that location.
-  void decodeAndSelectPlace(String placeId) {
+  void decodeAndSelectPlace(String? placeId) {
     clearOverlay();
 
     String endpoint =
@@ -242,7 +242,7 @@ class LocationPickerState extends State<LocationPicker> {
             '&language=${widget.language}';
 
     LocationUtils.getAppHeaders()
-        .then((headers) => http.get(Uri.parse(endpoint), headers: headers))
+        .then((headers) => http.get(Uri.parse(endpoint), headers: headers as Map<String, String>?))
         .then((response) {
       if (response.statusCode == 200) {
         Map<String, dynamic> location =
@@ -259,17 +259,17 @@ class LocationPickerState extends State<LocationPicker> {
 
   /// Display autocomplete suggestions with the overlay.
   void displayAutoCompleteSuggestions(List<RichSuggestion> suggestions) {
-    final RenderBox renderBox = context.findRenderObject();
+    final RenderBox renderBox = context.findRenderObject() as RenderBox;
     Size size = renderBox.size;
 
-    final RenderBox appBarBox = appBarKey.currentContext.findRenderObject();
+    final RenderBox? appBarBox = appBarKey.currentContext!.findRenderObject() as RenderBox?;
 
     clearOverlay();
 
     overlayEntry = OverlayEntry(
       builder: (context) => Positioned(
         width: size.width,
-        top: appBarBox.size.height,
+        top: appBarBox!.size.height,
         child: Padding(
           padding: const EdgeInsets.only(left: 16.0, right: 16.0),
           child: Material(
@@ -282,7 +282,7 @@ class LocationPickerState extends State<LocationPicker> {
       ),
     );
 
-    Overlay.of(context).insert(overlayEntry);
+    Overlay.of(context)!.insert(overlayEntry!);
   }
 
   /// Utility function to get clean readable name of a location. First checks
@@ -314,7 +314,7 @@ class LocationPickerState extends State<LocationPicker> {
                     "key=${widget.apiKey}&" +
                     "location=${latLng.latitude},${latLng.longitude}&radius=150" +
                     "&language=${widget.language}"),
-            headers: headers))
+            headers: headers as Map<String, String>?))
         .then((response) {
       if (response.statusCode == 200) {
         nearbyPlaces.clear();
@@ -351,13 +351,13 @@ class LocationPickerState extends State<LocationPicker> {
             "https://maps.googleapis.com/maps/api/geocode/json?latlng=${latLng.latitude},${latLng.longitude}" +
                 "&key=${widget.apiKey}" +
                 "&language=${widget.language}"),
-        headers: await LocationUtils.getAppHeaders());
+        headers: await (LocationUtils.getAppHeaders() as FutureOr<Map<String, String>?>));
 
     if (response.statusCode == 200) {
       Map<String, dynamic> responseJson = jsonDecode(response.body);
-      String road;
+      String? road;
 
-      String placeId = responseJson['results'][0]['place_id'];
+      String? placeId = responseJson['results'][0]['place_id'];
 
       if (responseJson['status'] == 'REQUEST_DENIED') {
         road = 'REQUEST DENIED = please see log for more details';
@@ -382,12 +382,12 @@ class LocationPickerState extends State<LocationPicker> {
   /// Moves the camera to the provided location and updates other UI features to
   /// match the location.
   void moveToLocation(LatLng latLng) {
-    mapKey.currentState.mapController.future.then((controller) {
+    mapKey.currentState!.mapController.future.then((controller) {
       controller.animateCamera(
         CameraUpdate.newCameraPosition(
           CameraPosition(
             target: latLng,
-            zoom: widget.initialZoom,
+            zoom: widget.initialZoom!,
           ),
         ),
       );
@@ -475,31 +475,31 @@ class LocationPickerState extends State<LocationPicker> {
 /// set [automaticallyAnimateToCurrentLocation] to false.
 ///
 ///
-Future<LocationResult> showLocationPicker(
+Future<LocationResult?> showLocationPicker(
   BuildContext context,
   String apiKey, {
   LatLng initialCenter = const LatLng(45.521563, -122.677433),
   double initialZoom = 16,
   bool requiredGPS = false,
-  List<String> countries,
+  List<String>? countries,
   bool myLocationButtonEnabled = false,
   bool layersButtonEnabled = false,
   bool automaticallyAnimateToCurrentLocation = true,
-  String mapStylePath,
+  String? mapStylePath,
   Color appBarColor = Colors.transparent,
-  Color layersIconColor,
-  Color layersButtonColor,
-  Color myLocationIconColor,
-  Color myLocationButtonColor,
-  Color selectButtonColor,
-  Color selectButtonFontColor,
-  String selectButtonText,
-  BoxDecoration searchBarBoxDecoration,
-  String hintText,
-  Widget resultCardConfirmIcon,
-  AlignmentGeometry resultCardAlignment,
-  EdgeInsetsGeometry resultCardPadding,
-  Decoration resultCardDecoration,
+  Color? layersIconColor,
+  Color? layersButtonColor,
+  Color? myLocationIconColor,
+  Color? myLocationButtonColor,
+  Color? selectButtonColor,
+  Color? selectButtonFontColor,
+  String? selectButtonText,
+  BoxDecoration? searchBarBoxDecoration,
+  String? hintText,
+  Widget? resultCardConfirmIcon,
+  AlignmentGeometry? resultCardAlignment,
+  EdgeInsetsGeometry? resultCardPadding,
+  Decoration? resultCardDecoration,
   String language = 'en',
   LocationAccuracy desiredAccuracy = LocationAccuracy.best,
 }) async {
@@ -528,8 +528,8 @@ Future<LocationResult> showLocationPicker(
           hintText: hintText,
           searchBarBoxDecoration: searchBarBoxDecoration,
           resultCardConfirmIcon: resultCardConfirmIcon,
-          resultCardAlignment: resultCardAlignment,
-          resultCardPadding: resultCardPadding,
+          resultCardAlignment: resultCardAlignment as Alignment?,
+          resultCardPadding: resultCardPadding as EdgeInsets?,
           resultCardDecoration: resultCardDecoration,
           countries: countries,
           language: language,
