@@ -5,8 +5,8 @@ import 'package:flutter/services.dart';
 
 class LocationUtils {
   static const _platform = const MethodChannel('google_map_location_picker');
-  static Map<String, String> _appHeaderCache = {};
-  static Future<Map<String, String>> getAppHeaders() async {
+  static Map<String, String?> _appHeaderCache = {};
+  static Future<Map<String, String?>> getAppHeaders() async {
     if (_appHeaderCache.isEmpty)
     {
       PackageInfo packageInfo = await PackageInfo.fromPlatform();
@@ -17,7 +17,7 @@ class LocationUtils {
           };
       }
       else if (Platform.isAndroid) {
-        String sha1;
+        String? sha1;
         try {
           sha1 = await _platform.invokeMethod(
               'getSigningCertSha1', packageInfo.packageName);
