@@ -13,7 +13,7 @@ import 'intl/messages_all.dart';
 class S {
   S();
   
-  static S current;
+  static late S current;
   
   static const AppLocalizationDelegate delegate =
     AppLocalizationDelegate();
@@ -29,7 +29,7 @@ class S {
     });
   } 
 
-  static S of(BuildContext context) {
+  static S? of(BuildContext context) {
     return Localizations.of<S>(context, S);
   }
 
